Remap Locales + Remove BasicTranslationService
